+++ test/compiler/compile.spec.js
@@ -9,10 +9,12 @@ describe('precompile', () => {
 
   it('Extension attribute', () => {
     const ret = precompile(`
-      <input type="radio" #show={{show}} #noMargin={{show}}>
+      <div type="radio" #show={{show}} #noMargin={{show}}>
+        <i><#test /></i>
+      </div>
     `, true, nj.tmplRule);
 
-    console.log(ret.main.toString());
+    //console.log(ret.main.toString());
     expect(Object.keys(ret).length).not.toBe(8);
     expect(ret.main.toString()).toContain('p1.x[\'noMargin\']');
   });